refactor : 경위도 검증 로직 이동 및 파일 위치 리펙토링
