+++ src/main/java/org/e2cho/e2cho_HW/service/externalApi/DataGoKrApiService.java
@@ -1,14 +1,14 @@
-package org.e2cho.e2cho_HW.service.weather;
+package org.e2cho.e2cho_HW.service.externalApi;
 
 import lombok.RequiredArgsConstructor;
 import org.e2cho.e2cho_HW.config.DataGoKrClient;
 import org.e2cho.e2cho_HW.config.DataGoKrClientConfig;
-import org.e2cho.e2cho_HW.dto.weather.DataGoKr;
+import org.e2cho.e2cho_HW.dto.externalApi.DataGoKr;
 import org.springframework.stereotype.Service;
 
 @Service
 @RequiredArgsConstructor
-public class DataGoKrService {
+public class DataGoKrApiService {
 
     private final DataGoKrClient dataGoKrClient;
     private final DataGoKrClientConfig dataGoKrClientConfig;
