+++ src/main/java/org/e2cho/e2cho_HW/service/externalApi/OpenWeatherMapApiService.java
@@ -1,14 +1,14 @@
-package org.e2cho.e2cho_HW.service.weather;
+package org.e2cho.e2cho_HW.service.externalApi;
 
 import lombok.RequiredArgsConstructor;
 import org.e2cho.e2cho_HW.config.OpenWeatherMapClient;
 import org.e2cho.e2cho_HW.config.OpenWeatherMapClientConfig;
-import org.e2cho.e2cho_HW.dto.weather.OpenWeatherMap;
+import org.e2cho.e2cho_HW.dto.externalApi.OpenWeatherMap;
 import org.springframework.stereotype.Service;
 
 @Service
 @RequiredArgsConstructor
-public class OpenWeatherMapService {
+public class OpenWeatherMapApiService {
 
     private final OpenWeatherMapClient openWeatherMapClient;
     private final OpenWeatherMapClientConfig openWeatherMapClientConfig;
