+++ src/main/java/org/e2cho/e2cho_HW/service/user/UserLocationService.java
@@ -9,6 +9,7 @@ import org.e2cho.e2cho_HW.dto.user.Location;
 import org.e2cho.e2cho_HW.exception.CustomErrorException;
 import org.e2cho.e2cho_HW.repository.user.UserLocationRepository;
 import org.e2cho.e2cho_HW.repository.user.UserRepository;
+import org.e2cho.e2cho_HW.service.util.LocationUtilService;
 import org.e2cho.e2cho_HW.service.util.UserUtilService;
 import org.springframework.stereotype.Service;
 
@@ -19,10 +20,10 @@ import static org.e2cho.e2cho_HW.constant.LocationBoundary.*;
 @RequiredArgsConstructor
 public class UserLocationService {
 
-    private final UserRepository userRepository;
     private final UserLocationRepository userLocationRepository;
 
     private final UserUtilService userUtilService;
+    private final LocationUtilService locationUtilService;
 
     @Transactional
     public Location.Dto saveUserLocation(Long userId, Location.Request request){
@@ -31,11 +32,7 @@ public class UserLocationService {
         User foundUser = userUtilService.getValidateUser(userId);
 
         // 2. request 로 받은 경위도 검사
-        if (!(EXTREME_SOUTH.getCoordinate() < request.getLatitude() && request.getLatitude() < EXTREME_NORTH.getCoordinate())
-                || !(EXTREME_WEST.getCoordinate() < request.getLongitude() && request.getLongitude() < EXTREME_EAST.getCoordinate())){
-
-            throw new CustomErrorException(ErrorType.OutOfBoundaryError);
-        }
+        locationUtilService.validateLocation(request.getLatitude(), request.getLongitude());
 
         // 3. 유저의 위치 정보가 이미 있다면, 경, 위도, 업데이트 시간 업데이트 없으면 처음 저장 로직 실행.
         UserLocation foundUserLocation = userLocationRepository.findByUserId(foundUser.getId());
